[APPC 192] Updated logic to send campaign id everytime a new proof cycle starts. Fix the reset of the proofs sent when a process is finished.
