+++ appcoins-ads/src/main/java/com/asf/appcoins/sdk/ads/AppCoinsAdsBuilder.java
@@ -2,7 +2,6 @@ package com.asf.appcoins.sdk.ads;
 
 import android.content.Context;
 import android.util.Log;
-import com.asf.appcoins.sdk.ads.campaign.manager.CampaignManager;
 import com.asf.appcoins.sdk.ads.ip.IpApi;
 import com.asf.appcoins.sdk.ads.ip.IpResponse;
 import com.asf.appcoins.sdk.ads.poa.PoAServiceConnector;
@@ -51,7 +50,7 @@ public final class AppCoinsAdsBuilder {
     if (debug) {
       networkId = NETWORK_ROPSTEN;
       web3 = Web3jFactory.build(new HttpService("https://ropsten.infura.io/1YsvKO0VH5aBopMYJzcy"));
-      contractAddress = new Address("0xd95c64c6eee9164539d679354f349779a04f57cb");
+      contractAddress = new Address("0xe8f3c5d390e3dfe1990c7ac5968f9aa5c5c808cd");
     } else {
       networkId = NETWORK_MAIN;
       web3 = Web3jFactory.build(new HttpService("https://mainnet.infura.io/1YsvKO0VH5aBopMYJzcy"));
@@ -71,7 +70,6 @@ public final class AppCoinsAdsBuilder {
         .onErrorReturn(throwable -> "")
         .blockingFirst();
 
-    return new AppCoinsAdsImpl(poaConnector, networkId,
-        new CampaignManager(asfWeb3j, contractAddress, countryId));
+    return new AppCoinsAdsImpl(poaConnector, networkId, asfWeb3j, contractAddress, countryId);
   }
 }
